rename fuse/index to fuse/pruning
