+++ query/src/storages/fuse/io/locations.rs
