+++ query/src/storages/fuse/io/mod.rs
@@ -14,12 +14,12 @@
 //
 
 #[cfg(test)]
-mod block_appender_test;
+mod block_stream_writer_test;
 
-mod block_appender;
-mod col_encoding;
-mod location_gen;
+mod block_stream_writer;
+mod block_writer;
+mod locations;
 
-pub use block_appender::BlockAppender;
-pub use location_gen::gen_segment_info_location;
-pub use location_gen::snapshot_location;
+pub use block_stream_writer::BlockStreamWriter;
+pub use locations::gen_segment_info_location;
+pub use locations::snapshot_location;
