+++ query/src/storages/fuse/operations/read_plan.rs
@@ -24,8 +24,8 @@ use common_planners::Partitions;
 use common_planners::Statistics;
 
 use crate::sessions::QueryContext;
-use crate::storages::fuse::index::apply_range_filter;
 use crate::storages::fuse::meta::BlockMeta;
+use crate::storages::fuse::pruning::apply_block_pruning;
 use crate::storages::fuse::FuseTable;
 
 impl FuseTable {
@@ -41,7 +41,7 @@ impl FuseTable {
             let schema = self.table_info.schema();
             let push_downs_c = push_downs.clone();
             let snapshot = read_obj(da.clone(), loc).await?;
-            let block_metas = apply_range_filter(&snapshot, schema, push_downs_c, da).await?;
+            let block_metas = apply_block_pruning(&snapshot, schema, push_downs_c, da).await?;
             let (statistics, parts) = Self::to_partitions(&block_metas, push_downs);
             Ok((statistics, parts))
         } else {
