+++ query/src/storages/fuse/pruning/block_pruner_test.rs
@@ -30,13 +30,13 @@ use common_planners::Extras;
 use futures::TryStreamExt;
 
 use crate::catalogs::Catalog;
-use crate::storages::fuse::index::min_max::apply_range_filter;
+use crate::storages::fuse::pruning::block_pruner::apply_block_pruning;
 use crate::storages::fuse::table_test_fixture::TestFixture;
 use crate::storages::fuse::TBL_OPT_KEY_CHUNK_BLOCK_NUM;
 use crate::storages::fuse::TBL_OPT_KEY_SNAPSHOT_LOC;
 
 #[tokio::test]
-async fn test_min_max_index() -> Result<()> {
+async fn test_block_pruner() -> Result<()> {
     let fixture = TestFixture::new().await;
     let ctx = fixture.ctx();
 
@@ -100,7 +100,7 @@ async fn test_min_max_index() -> Result<()> {
 
     // no pruning
     let push_downs = None;
-    let blocks = apply_range_filter(
+    let blocks = apply_block_pruning(
         &snapshot,
         table.get_table_info().schema(),
         push_downs,
@@ -116,7 +116,7 @@ async fn test_min_max_index() -> Result<()> {
     let pred = col("a").gt(lit(30));
     extra.filters = vec![pred];
 
-    let blocks = apply_range_filter(
+    let blocks = apply_block_pruning(
         &snapshot,
         table.get_table_info().schema(),
         Some(extra),
@@ -131,7 +131,7 @@ async fn test_min_max_index() -> Result<()> {
     extra.filters = vec![pred];
 
     let blocks =
-        apply_range_filter(&snapshot, table.get_table_info().schema(), Some(extra), da).await?;
+        apply_block_pruning(&snapshot, table.get_table_info().schema(), Some(extra), da).await?;
     assert_eq!(num - 1, blocks.len() as u64);
 
     Ok(())
